Rename directory, update README
