+++ pyproject.toml
@@ -3,7 +3,7 @@
     build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-    packages = ["src"]
+    packages = ["coupling"]
 
 [project]
     name = "richards-swe"
